FastApi/requirements.txt
aiofiles==0.6.0
APScheduler==3.6.3
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.4
click==7.1.2
colorama==0.4.4
cryptography==3.3.1
dnspython==2.0.0
ecdsa==0.14.1
elasticsearch==7.10.1
email-validator==1.1.2
fastapi==0.63.0
h11==0.11.0
idna==2.10
Jinja2==2.11.2
loguru==0.5.3
MarkupSafe==1.1.1
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.20
pydantic==1.7.3
PyMySQL==0.10.1
python-jose==3.2.0
python-multipart==0.0.5
pytz==2020.5
rsa==4.6
six==1.15.0
SQLAlchemy==1.3.22
starlette==0.13.6
tzlocal==2.1
urllib3==1.26.2
uvicorn==0.13.2
win32-setctime==1.0.3
